Strengthen Python() tests (#2586)
* Strengthen Python() tests

* Verify raw vs parsed

* Verify injection

* Rename test_versioning/ to versioning/

* Split fixtures into conftest

* Split Python() tests

* Test raw constraints

* Consolidate raw in repr
