+++ client/verta/tests/versioning/test_configuration.py
